Add authenticate step to necessary routes

+++ src/main/java/com/github/enteraname74/musik/controller/PlaylistController.java
@@ -1,22 +1,26 @@
 package com.github.enteraname74.musik.controller;
 
+import com.github.enteraname74.musik.controller.utils.ControllerUtils;
 import com.github.enteraname74.musik.domain.model.Playlist;
+import com.github.enteraname74.musik.domain.service.AuthService;
 import com.github.enteraname74.musik.domain.service.PlaylistService;
 import com.github.enteraname74.musik.domain.utils.ServiceResult;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
 
 @RestController
 @RequestMapping("/playlist")
 public class PlaylistController {
     private final PlaylistService playlistService;
+    private final AuthService authService;
 
     @Autowired
-    public PlaylistController(PlaylistService playlistService) {
+    public PlaylistController(PlaylistService playlistService, AuthService authService) {
         this.playlistService = playlistService;
+        this.authService = authService;
     }
 
     /**
@@ -27,8 +31,10 @@ public class PlaylistController {
      */
     @PostMapping()
     public ResponseEntity<?> add(
+            @RequestHeader(name = HttpHeaders.AUTHORIZATION) String token,
             @RequestBody Playlist playlist
     ) {
+        if (!authService.isUserAuthenticated(token)) return ControllerUtils.UNAUTHORIZED_RESPONSE;
         ServiceResult<?> result =  playlistService.save(playlist);
 
         return new ResponseEntity<>(result.getResult(), result.getHttpStatus());
@@ -40,7 +46,16 @@ public class PlaylistController {
      * @return a list containing all Playlists or an empty list if there is an error.
      */
     @GetMapping("/all")
-    List<Playlist> getAll() { return playlistService.getAll(); }
+    ResponseEntity<?> getAll(
+            @RequestHeader(name = HttpHeaders.AUTHORIZATION) String token
+    ) {
+        if (!authService.isUserAuthenticated(token)) return ControllerUtils.UNAUTHORIZED_RESPONSE;
+
+        return new ResponseEntity<>(
+                playlistService.getAll(),
+                HttpStatus.OK
+        );
+    }
 
     /**
      * Retrieves a Playlist from its id.
@@ -50,8 +65,11 @@ public class PlaylistController {
      */
     @GetMapping("/{id}")
     ResponseEntity<?> get(
+            @RequestHeader(name = HttpHeaders.AUTHORIZATION) String token,
             @PathVariable String id
     ) {
+        if (!authService.isUserAuthenticated(token)) return ControllerUtils.UNAUTHORIZED_RESPONSE;
+
         ServiceResult<?> result = playlistService.getById(id);
 
         return new ResponseEntity<>(result.getResult(), result.getHttpStatus());
@@ -66,8 +84,10 @@ public class PlaylistController {
      */
     @DeleteMapping("/{id}")
     ResponseEntity<?> delete(
+            @RequestHeader(name = HttpHeaders.AUTHORIZATION) String token,
             @PathVariable String id
     ) {
+        if (!authService.isUserAuthenticated(token)) return ControllerUtils.UNAUTHORIZED_RESPONSE;
         ServiceResult<?> result = playlistService.deleteById(id);
 
         return new ResponseEntity<>(result.getResult(), result.getHttpStatus());
@@ -82,9 +102,11 @@ public class PlaylistController {
      */
     @GetMapping("/add/{playlistId}/{musicId}")
     ResponseEntity<?> addMusicToPlaylist(
+            @RequestHeader(name = HttpHeaders.AUTHORIZATION) String token,
             @PathVariable String playlistId,
             @PathVariable String musicId
     ) {
+        if (!authService.isUserAuthenticated(token)) return ControllerUtils.UNAUTHORIZED_RESPONSE;
         ServiceResult<?> result = playlistService.addMusicToPlaylist(playlistId, musicId);
 
         return new ResponseEntity<>(result.getResult(), result.getHttpStatus());
@@ -99,9 +121,11 @@ public class PlaylistController {
      */
     @GetMapping("/remove/{playlistId}/{musicId}")
     ResponseEntity<?> removeMusicFromPlaylist(
+            @RequestHeader(name = HttpHeaders.AUTHORIZATION) String token,
             @PathVariable String playlistId,
             @PathVariable String musicId
     ) {
+        if (!authService.isUserAuthenticated(token)) return ControllerUtils.UNAUTHORIZED_RESPONSE;
         ServiceResult<?> result = playlistService.removeMusicToPlaylist(playlistId, musicId);
 
         return new ResponseEntity<>(result.getResult(), result.getHttpStatus());
